Removed all skill flags, updated utility, added skill vector functions
- Updated utility to use vectors
- Updated RVC format completely, removing all references to version 1
- Removed addSpells, moving it to the modClass function
- Devised function in Character::createNewCharacter to list all usable skills and store in the vector in Character, leaving the class vector alone with only class skills
- other bug fixes

+++ Spell.cpp
@@ -1,6 +1,6 @@
 #include "Spell.h"
 
-uint8_t Spell::getLevel() {
+short Spell::getLevel() {
 	return spell_level;
 }
